rename log_setting to _logging, status_monitor to _status_monitor
