+++ src/otaclient/create_standby/common.py
@@ -34,7 +34,7 @@ from weakref import WeakKeyDictionary, WeakValueDictionary
 from ota_metadata.legacy.parser import MetafilesV1, OTAMetadata
 from ota_metadata.legacy.types import DirectoryInf, RegularInf
 from otaclient.configs.cfg import cfg
-from otaclient.status_monitor import StatusReport, UpdateProgressReport
+from otaclient._status_monitor import StatusReport, UpdateProgressReport
 from otaclient_common.common import create_tmp_fname
 
 logger = logging.getLogger(__name__)
