+++ src/otaclient/grpc/api_v2/ecu_tracker.py
@@ -22,7 +22,7 @@ import logging
 from otaclient.configs import ECUContact
 from otaclient.configs.cfg import cfg, ecu_info
 from otaclient.grpc.api_v2.ecu_status import ECUStatusStorage
-from otaclient.status_monitor import OTAClientStatusCollector
+from otaclient._status_monitor import OTAClientStatusCollector
 from otaclient_api.v2 import types as api_types
 from otaclient_api.v2.api_caller import ECUNoResponse, OTAClientCall
