+++ tests/test_otaclient/test_status_monitor.py
@@ -26,7 +26,7 @@ from typing import Generator
 import pytest
 
 from otaclient._types import FailureType, OTAStatus, UpdatePhase
-from otaclient.status_monitor import (
+from otaclient._status_monitor import (
     TERMINATE_SENTINEL,
     OTAClientStatusCollector,
     OTAStatusChangeReport,
